Move libs to lib/
Cherry-picked from https://github.com/GooborgStudios/mdn-bcd-collector/commit/08cc5db05eb39e98dd769ecb7fc5477538bcaecc.
