+++ result-stats.ts
@@ -17,8 +17,8 @@ import {hideBin} from 'yargs/helpers';
 import {CompatData} from '@mdn/browser-compat-data/types';
 
 import {Report} from './types/types.js';
+import {parseUA} from './lib/ua-parser.js';
 import {findMissing} from './find-missing-features.js';
-import {parseUA} from './ua-parser.js';
 
 const BCD_DIR = process.env.BCD_DIR || `../browser-compat-data`;
 const {default: bcd}: {default: CompatData} = await import(
